requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
flask_login==0.6.3
flask_socketio==5.3.6
python-dotenv==1.0.1
flask_wtf==1.2.1
email_validator==2.1.1
unidecode==1.3.8
flask_bcrypt==1.0.1
Whoosh==2.7.4
spacy==3.7.4
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
pandas==2.2.2
six==1.16.0
pytesseract==0.3.10
xlrd==2.0.1
pymupdf==1.24.1
openpyxl==3.1.2
chardet==5.2.0
python-pptx==0.6.23
python-docx==1.1.0
pdf2image==1.17.0
celery==5.3.6
fasteners==0.19
flask_redis==0.4.0
fr_hades_model @ https://github.com/Nathan-Etave/Hades/releases/download/fr_hades_model-1.3.0/fr_hades_model-1.3.0-py3-none-any.whl
gunicorn==22.0.0